Moved role change tests to separate test file
